remove get_all_blob js implementation
