+++ get_all_blobs/.gitignore
@@ -1,2 +1,1 @@
-eigenda-proxy/
 blob_data.json
