--- get_all_blobs/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "get_all_blobs",
-  "version": "1.0.0",
-  "description": "",
-  "main": "getallblobs.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "web3": "^4.12.1"
-  }
-}
