+++ get_all_blobs/src/blob_info.rs
